requirements.txt
gitpython
filetype
tkextrafont
watchdog
pygments
Pillow
toml
hintedtext
sv-ttk
requests
pyperclip
tkinterweb
mistune
tarts
chardet 
python-lsp-server
pywinpty; platform_system == "Windows"
ptyprocess; platform_system != "Windows"
python-tkdnd
psutil
google-generativeai
click